src/test/java/hello/hellospring/repository/MemoryMemberRepositoryTest.java
package hello.hellospring.repository;

import hello.hellospring.domain.Member;
import org.junit.jupiter.api.AfterEach;
import org.junit.jupiter.api.Test;


import java.util.List;

import static org.assertj.core.api.AssertionsForClassTypes.assertThat;


class MemoryMemberRepositoryTest {

    MemoryMemberRepository repository = new MemoryMemberRepository();

    @AfterEach
    public void afterEach() {//콜백메서드
        repository.clearStore();
    }

    @Test
    public void sava() {
        Member member = new Member();
        member.setName("spring");

        repository.sava(member);

        Member result = repository.findById(member.getId()).get();
        //  System.out.println("result = " + (result == member));
        //Assertions.assertEquals(member, result);
        assertThat(member).isEqualTo(result);
    }

    @Test
    public void findByName() {
        Member member1 = new Member();
        member1.setName("spring1");
        repository.sava(member1);

        Member member2 = new Member();
        member2.setName("spring2");
        repository.sava(member2);

        Member result = repository.findByName("spring1").get();

        assertThat(result).isEqualTo(member1);
    }

    @Test
    public void findAll() {
        Member member1 = new Member();
        member1.setName("spring1");
        repository.sava(member1);

        Member member2 = new Member();
        member2.setName("spring2");
        repository.sava(member2);

        List<Member> result = repository.findAll();

        assertThat(result.size()).isEqualTo(2);
    }
}